refactor: split tests into separate module file
